extract module to file
